requirements.txt
# --- Core Framework & Server ---
Django
django
djangorestframework
daphne  # ASGI server for future real-time features with Channels

# --- Database ---
psycopg2-binary   # PostgreSQL driver
dj-database-url   # For parsing DATABASE_URL from .env
pgvector   # For vector search capabilities in PostgreSQL

# --- Asynchronous & Background Tasks ---
celery
redis
django_redis
django-celery-beat  # For scheduling our periodic tasks
django-celery-results
httpx             # Modern, async-first HTTP client

# --- AI & Language Models ---
openai              # For embeddings and/or chat models
google-generativeai # For using the Gemini API

# --- Web Scraping ---
beautifulsoup4
requests          # A fallback/synchronous HTTP client

# --- Environment & Utilities ---
environs
python-decouple     # For cleanly managing .env variables


# --- Other Third Part-Packages  ---
whitenoise
django-cors-headers



